fix: replace AVAudioUnitDynamicsProcessor + 2 type errors
- AutoMixChain: replace deprecated AVAudioUnitDynamicsProcessor (compressor/limiter)
  with AVAudioMixerNode gainNode — works on all platforms including macCatalyst
- ClipEngine: voice.harmonicity = Double(scene.harmonicity) → scene.harmonicity (Float)
- MasterView: conf > Float(i) * 0.25 → Double(i) (conf is Double, no implicit cast)

+++ Sources/Echoelmusic/Views/MasterView.swift
@@ -845,7 +845,7 @@ struct MasterView: View {
             HStack(spacing: 3) {
                 ForEach(0..<4, id: \.self) { i in
                     RoundedRectangle(cornerRadius: 1)
-                        .fill(Color.white.opacity(conf > Float(i) * 0.25 ? 0.28 : 0.05))
+                        .fill(Color.white.opacity(conf > Double(i) * 0.25 ? 0.28 : 0.05))
                         .frame(width: 12, height: 3)
                 }
             }
